+++ CHANGELOG.md
@@ -2,6 +2,8 @@ divelist: do not include planned versions of a dive if there is real data
 desktop: fix key composition in tag widgets and dive site widget
 mobile: allow cloud account deletion (Apple app store requirement)
 mobile: fix listing of local cloud cache directories
+dive computer support:
+- fix Ratio dive computer detection for BLE only devices
 
 ---
 * Always add new entries at the very top of this file above other existing entries and this note.